Removed hardcoding of augmentation and rescaling.
